@@ -1,22 +1,34 @@
 {
-  "name": "har-extractor",
-  "version": "1.1.2",
-  "description": "A CLI that extract har file to directory.",
+  "name": "har-extractor-easy",
+  "version": "1.0.1",
+  "type": "module",
+  "description": "A (easy) CLI that extract har file to unique files in a directory.",
+  "scripts": {
+    "build": "cross-env NODE_ENV=production tsc -p .",
+    "clean": "rimraf lib/",
+    "prepublish": "npm run --if-present build",
+    "prettier": "prettier --write \"**/*.{js,jsx,ts,tsx,css}\"",
+    "watch": "tsc -p . --watch",
+    "prepare": "npm run build",
+    "version": "npm run prettier && git add -A src",
+    "postversion": "git push && git push --tags"
+  },
   "keywords": [
-    "archieve",
-    "directory",
-    "extract",
-    "file",
-    "folder",
     "har",
+    "extract",
+    "decode",
+    "query",
+    "files",
+    "folder",
+    "chrome",
     "unzip"
   ],
-  "homepage": "https://github.com/azu/har-extractor",
+  "homepage": "https://github.com/aceslick911/har-extractor-easy#readme",
   "bugs": {
-    "url": "https://github.com/azu/har-extractor/issues"
+    "url": "https://github.com/aceslick911/har-extractor-easy/issues"
   },
   "license": "MIT",
-  "author": "azu",
+  "author": "aceslick911",
   "files": [
     "bin/",
     "lib/",
@@ -28,20 +40,11 @@
     "har-extractor": "./bin/cmd.js"
   },
   "directories": {
-    "lib": "lib",
-    "test": "test"
+    "lib": "lib"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/azu/har-extractor.git"
-  },
-  "scripts": {
-    "build": "cross-env NODE_ENV=production tsc -p .",
-    "clean": "rimraf lib/",
-    "prepublish": "npm run --if-present build",
-    "test": "mocha \"test/**/*.ts\"",
-    "prettier": "prettier --write \"**/*.{js,jsx,ts,tsx,css}\"",
-    "watch": "tsc -p . --watch"
+    "url": "https://github.com/aceslick911/har-extractor-easy.git"
   },
   "prettier": {
     "printWidth": 120,
@@ -50,19 +53,15 @@
   },
   "devDependencies": {
     "@types/har-format": "^1.2.5",
-    "@types/mocha": "^8.2.1",
+    "@types/mime-db": "^1.43.1",
     "@types/node": "^14.14.35",
     "cross-env": "^7.0.3",
     "del": "^6.0.0",
     "glob": "^7.1.6",
-    "husky": "^3.1.0",
     "lint-staged": "^9.5.0",
-    "mocha": "^8.3.2",
     "prettier": "^2.2.1",
     "rimraf": "^3.0.2",
-    "ts-node": "^9.1.1",
-    "ts-node-test-register": "^9.0.1",
-    "typescript": "^4.2.3"
+    "typescript": "^4.8.3"
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx,css}": [
@@ -71,15 +70,12 @@
     ]
   },
   "dependencies": {
+    "@types/humanize-url": "^2.1.0",
     "filenamify": "^2.1.0",
     "humanize-url": "^1.0.1",
     "make-dir": "^1.3.0",
-    "meow": "^5.0.0"
-  },
-  "husky": {
-    "hooks": {
-      "post-commit": "git reset",
-      "pre-commit": "lint-staged"
-    }
+    "meow": "^5.0.0",
+    "mime-db": "^1.52.0",
+    "unused-filename": "^4.0.1"
   }
 }
