@@ -1,55 +1,36 @@
 import * as fs from "fs";
-import { Har, Entry } from "har-format";
+import { Har } from "har-format";
 import * as path from "path";
 
-const filenamify = require("filenamify");
-const humanizeUrl = require("humanize-url");
-const makeDir = require("make-dir");
-export const getEntryContentAsBuffer = (entry: Entry): Buffer | undefined => {
-    const content = entry.response.content;
-    const text = content.text;
-    if (text === undefined) {
-        return;
-    }
-    if (content.encoding === "base64") {
-        return Buffer.from(text, "base64");
-    } else {
-        return Buffer.from(text);
-    }
-};
+//@ts-ignore
+import makeDir from "make-dir";
+import getEntryContentAsBuffer from "./data-utils.js";
 
-export const convertEntryAsFilePathFormat = (entry: Entry, removeQueryString: boolean = false): string => {
-    const requestURL = entry.request.url;
-    const stripSchemaURL: string = humanizeUrl(removeQueryString ? requestURL.split("?")[0] : requestURL);
-    const dirnames: string[] = stripSchemaURL.split("/").map((pathname) => {
-        return filenamify(pathname, {maxLength: 255});
-    });
-    const fileName = dirnames[dirnames.length - 1];
-    if (
-        fileName &&
-        !fileName.includes(".html") &&
-        entry.response.content.mimeType &&
-        entry.response.content.mimeType.includes("text/html")
-    ) {
-        return dirnames.join("/") + "/index.html";
-    }
-    return dirnames.join("/");
-};
+import { convertEntryAsFilePathFormat, getAvailableFilename } from "./filename-resolve.js";
 
 export interface ExtractOptions {
     outputDir: string;
     verbose?: boolean;
     dryRun?: boolean;
     removeQueryString?: boolean;
+    pretty?: boolean;
 }
 
 export const extract = (harContent: Har, options: ExtractOptions) => {
-    harContent.log.entries.forEach((entry) => {
+    harContent.log.entries.forEach(async (entry) => {
         const buffer = getEntryContentAsBuffer(entry);
         if (!buffer) {
             return;
         }
-        const outputPath = path.join(options.outputDir, convertEntryAsFilePathFormat(entry, options.removeQueryString));
+        const { uniquePath, updatedBuffer } = convertEntryAsFilePathFormat(
+            buffer,
+            entry,
+            options.removeQueryString,
+            options.pretty
+        );
+
+        const outputPath = getAvailableFilename(path.join(options.outputDir, uniquePath));
+
         if (!options.dryRun) {
             makeDir.sync(path.dirname(outputPath));
         }
@@ -57,7 +38,7 @@ export const extract = (harContent: Har, options: ExtractOptions) => {
             console.log(outputPath);
         }
         if (!options.dryRun) {
-            fs.writeFileSync(outputPath, buffer);
+            fs.writeFileSync(outputPath, updatedBuffer);
         }
     });
 };
